New convention in fset of model_object properties.
fset has an optional argument init that defaults to False.
When set to True, the property is set without calling functions
that might rely on other properties or without assuming that
a previous value of the property exists. This typically happens
when a tree of objects is loaded from a zml file. The new
implementation avoids that the attribute 'targets' is
created for a referent object. This list is not used and only
causes confusion. The second advantage is that internal attribute
names can change without changing the zml file format.